+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Update callables to use 'main' entrypoint instead of the old 'app' entrypoint.
 - Moved the watchPod implementation to the generic Watchers package.
 - Callable does not fail immediately if it does not find a service, but it retries.
+- On Grafana annotations print the extracted Kind of the object instead of the reflected type.
 - ...
 
 ### New Features & Functionality